moved array element to commons project
