--- module-array/src/main/java/ch/hsr/adv/ui/array/logic/domain/ArrayElement.java
@@ -1,63 +1,0 @@
-package ch.hsr.adv.ui.array.logic.domain;
-
-import ch.hsr.adv.commons.core.logic.domain.ADVElement;
-import ch.hsr.adv.commons.core.logic.domain.styles.ADVStyle;
-import ch.hsr.adv.commons.core.logic.domain.styles.presets.ADVDefaultStyle;
-
-/**
- * Represents one Item in a Array
- */
-public class ArrayElement implements ADVElement<String> {
-
-    private long id;
-    private int fixedPosX;
-    private int fixedPosY;
-    private String content;
-    private ADVStyle style = new ADVDefaultStyle();
-
-    @Override
-    public long getId() {
-        return id;
-    }
-
-    @Override
-    public ADVStyle getStyle() {
-        return style;
-    }
-
-    public void setStyle(ADVStyle style) {
-        this.style = style;
-    }
-
-    @Override
-    public int getFixedPosX() {
-        return fixedPosX;
-    }
-
-    public void setFixedPosX(int fixedPosX) {
-        this.fixedPosX = fixedPosX;
-    }
-
-    @Override
-    public int getFixedPosY() {
-        return fixedPosY;
-    }
-
-    public void setFixedPosY(int fixedPosY) {
-        this.fixedPosY = fixedPosY;
-    }
-
-    @Override
-    public String getContent() {
-        return content;
-    }
-
-    public void setContent(String content) {
-        this.content = content;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-}
